Remove Fishing Spot hotkey
This was mostly removed due to the mod being more focused on reporting bugs on MCC Island more easily. The Fishing Spot sharing hotkey was mostly a gameplay feature to help others in early testing. Since Fishing is no longer in a testing state, this feature will be removed and should be in another mod that focuses more on gameplay instead of reporting bugs, e.g. Island Utils (if they want to implement that feature).